# openapi_contract.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "URL Shortener",
    "version": "2.0.0"
  },
  "paths": {
...
    "/api/projects/{project_id}/links": {
      "post": {
        "tags": [
          "Links"
        ],
        "summary": "Create New Link",
        "operationId": "create_new_link_api_projects__project_id__links_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateLinkRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LinkResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Links"
        ],
        "summary": "List Project Links",
        "operationId": "list_project_links_api_projects__project_id__links_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "per_page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Per Page"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LinkListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/{project_id}/links/{link_id}": {
      "get": {
        "tags": [
          "Links"
        ],
        "summary": "Get Link",
        "operationId": "get_link_api_projects__project_id__links__link_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          },
          {
            "name": "link_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Link Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LinkResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Links"
        ],
        "summary": "Update Existing Link",
        "operationId": "update_existing_link_api_projects__project_id__links__link_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          },
          {
            "name": "link_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Link Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateLinkRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LinkResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Links"
        ],
        "summary": "Delete Existing Link",
        "operationId": "delete_existing_link_api_projects__project_id__links__link_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          },
          {
            "name": "link_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Link Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/{project_id}/links/{link_id}/analytics": {
      "get": {
        "tags": [
          "Links"
        ],
        "summary": "Get Link Stats",
        "operationId": "get_link_stats_api_projects__project_id__links__link_id__analytics_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          },
          {
            "name": "link_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Link Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LinkAnalyticsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/platform/links": {
      "get": {
        "tags": [
          "Platform Admin"
        ],
        "summary": "List All Links",
        "operationId": "list_all_links_api_platform_links_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "name": "per_page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Per Page"
            }
          },
          {
            "name": "project_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlatformLinkListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ClicksOverTime": {
        "properties": {
          "date": {
            "type": "string",
            "title": "Date"
          },
          "clicks": {
            "type": "integer",
            "title": "Clicks"
          }
        },
        "type": "object",
        "required": [
          "date",
          "clicks"
        ],
        "title": "ClicksOverTime"
      },
...
      "CreateLinkRequest": {
        "properties": {
          "original_url": {
            "type": "string",
            "maxLength": 2083,
            "minLength": 1,
            "format": "uri",
            "title": "Original Url"
          },
          "expiry": {
            "$ref": "#/components/schemas/ExpiryOption",
            "default": "never"
          },
          "custom_code": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 20,
                "minLength": 4
              },
              {
                "type": "null"
              }
            ],
            "title": "Custom Code"
          },
          "title": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          }
        },
        "type": "object",
        "required": [
          "original_url"
        ],
        "title": "CreateLinkRequest"
      },
...
      "ExpiryOption": {
        "type": "string",
        "enum": [
          "1d",
          "7d",
          "30d",
          "90d",
          "never"
        ],
        "title": "ExpiryOption"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "LinkAnalyticsResponse": {
        "properties": {
          "link_id": {
            "type": "string",
            "title": "Link Id"
          },
          "short_code": {
            "type": "string",
            "title": "Short Code"
          },
          "total_clicks": {
            "type": "integer",
            "title": "Total Clicks"
          },
          "unique_visitors": {
            "type": "integer",
            "title": "Unique Visitors"
          },
          "clicks_today": {
            "type": "integer",
            "title": "Clicks Today"
          },
          "clicks_this_week": {
            "type": "integer",
            "title": "Clicks This Week"
          },
          "clicks_this_month": {
            "type": "integer",
            "title": "Clicks This Month"
          },
          "devices": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Devices"
          },
          "browsers": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Browsers"
          },
          "os_breakdown": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Os Breakdown"
          },
          "referers": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Referers"
          },
          "clicks_over_time": {
            "items": {
              "$ref": "#/components/schemas/ClicksOverTime"
            },
            "type": "array",
            "title": "Clicks Over Time"
          },
          "peak_hour": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Peak Hour"
          },
          "bot_clicks": {
            "type": "integer",
            "title": "Bot Clicks"
          },
          "human_clicks": {
            "type": "integer",
            "title": "Human Clicks"
          }
        },
        "type": "object",
        "required": [
          "link_id",
          "short_code",
          "total_clicks",
          "unique_visitors",
          "clicks_today",
          "clicks_this_week",
          "clicks_this_month",
          "devices",
          "browsers",
          "os_breakdown",
          "referers",
          "clicks_over_time",
          "peak_hour",
          "bot_clicks",
          "human_clicks"
        ],
        "title": "LinkAnalyticsResponse"
      },
      "LinkListResponse": {
        "properties": {
          "links": {
            "items": {
              "$ref": "#/components/schemas/LinkResponse"
            },
            "type": "array",
            "title": "Links"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "per_page": {
            "type": "integer",
            "title": "Per Page"
          }
        },
        "type": "object",
        "required": [
          "links",
          "total",
          "page",
          "per_page"
        ],
        "title": "LinkListResponse"
      },
      "LinkResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "project_id": {
            "type": "string",
            "title": "Project Id"
          },
          "short_code": {
            "type": "string",
            "title": "Short Code"
          },
          "short_url": {
            "type": "string",
            "title": "Short Url"
          },
          "original_url": {
            "type": "string",
            "title": "Original Url"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "expires_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "click_count": {
            "type": "integer",
            "title": "Click Count"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "project_id",
          "short_code",
          "short_url",
          "original_url",
          "title",
          "expires_at",
          "is_active",
          "click_count",
          "created_at",
          "updated_at"
        ],
        "title": "LinkResponse"
      },
...
      "PlatformLinkListResponse": {
        "properties": {
          "links": {
            "items": {
              "$ref": "#/components/schemas/PlatformLinkResponse"
            },
            "type": "array",
            "title": "Links"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "type": "object",
        "required": [
          "links",
          "next_cursor"
        ],
        "title": "PlatformLinkListResponse"
      },
      "PlatformLinkResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "project_id": {
            "type": "string",
            "title": "Project Id"
          },
          "created_by": {
            "type": "string",
            "title": "Created By"
          },
          "short_code": {
            "type": "string",
            "title": "Short Code"
          },
          "original_url": {
            "type": "string",
            "title": "Original Url"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "click_count": {
            "type": "integer",
            "title": "Click Count"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "project_id",
          "created_by",
          "short_code",
          "original_url",
          "title",
          "is_active",
          "click_count",
          "created_at"
        ],
        "title": "PlatformLinkResponse"
      },
...
      "UpdateLinkRequest": {
        "properties": {
          "title": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "is_active": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Active"
          }
        },
        "type": "object",
        "title": "UpdateLinkRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}